refactor: 테스트 코드 public 제거
- 테스트 코드 메서드에서 public 제거
