--- src/main/java/racingcar/domain/race/Race.java
@@ -1,4 +1,0 @@
-package racingcar.domain.race;
-
-public class Race {
-}
