--- src/main/java/racingcar/domain/race/Raceable.java
@@ -1,4 +1,0 @@
-package racingcar.domain.race;
-
-public interface Raceable {
-}
