--- src/main/java/racingcar/domain/winner/Winnable.java
@@ -1,4 +1,0 @@
-package racingcar.domain.winner;
-
-public interface Winnable {
-}
