--- src/main/java/racingcar/domain/winner/Winner.java
@@ -1,4 +1,0 @@
-package racingcar.domain.winner;
-
-public class Winner {
-}
